+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: izaitcev <izaitcev@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/05/16 16:31:44 by izaitcev      #+#    #+#                 */
-/*   Updated: 2022/07/11 16:17:43 by izaitcev      ########   odam.nl         */
+/*   Updated: 2022/07/11 16:32:27 by izaitcev      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int	print_char(int c, size_t *len);
 int	print_nbr(int n);
 int	print_unsigned_dec(unsigned int n);
 int	print_lhex(unsigned long n);
-int	print_uphex(unsigned long n);
+int	print_uphex(unsigned int n);
 int	print_percent(int c);
 
 #endif
